refactor(go-template): extract prop-type resolution into props/prop-types.ts
Move the prop-type resolution helpers out of the adapter class into pure
free functions (Roadmap B): buildPropTypeOverrides, resolvePropGoType,
collectNillablePropNames. They read only state.localStructFields and
extractPropNameFromInitialValue and resolve via typeInfoToGo — no new
GoEmitContext member.

The struct-assembly generators (generateInputStruct / generatePropsStruct
/ generateNewPropsFunction) stay on the adapter as the orchestrator core
(~18 cross-method deps); the architecture keeps composition on the object
rather than re-exposing it through the seam.

Output byte-identical, verified by the adapter unit (552) and
adapter-tests conformance (786) suites; tsgo + biome clean.
